test: refactor ruby and php modules to use mocked commands  (#936)
